C23 - ORM

+++ src/main/java/com/dh/Backend1_TrabajoIntegrador/servicio/IOdontologoServicio.java
@@ -1,13 +1,13 @@
 package com.dh.Backend1_TrabajoIntegrador.servicio;
 
-import com.dh.Backend1_TrabajoIntegrador.modelo.Odontologo;
+import com.dh.Backend1_TrabajoIntegrador.entidad.Odontologo;
 
 import java.util.List;
 
 public interface IOdontologoServicio {
-    Odontologo consultarPorId(Integer id);
+    Odontologo consultarPorId(Long id);
     List<Odontologo> consultarTodos();
     Odontologo guardar(Odontologo odontologo);
     Odontologo modificar(Odontologo odontologo);
-    void eliminar(Integer id);
+    void eliminar(Long id);
 }
