+++ src/main/java/com/dh/Backend1_TrabajoIntegrador/servicio/IPacienteServicio.java
@@ -1,13 +1,13 @@
 package com.dh.Backend1_TrabajoIntegrador.servicio;
 
-import com.dh.Backend1_TrabajoIntegrador.modelo.Paciente;
+import com.dh.Backend1_TrabajoIntegrador.entidad.Paciente;
 
 import java.util.List;
 
 public interface IPacienteServicio {
-    Paciente consultarPorId(Integer id);
+    Paciente consultarPorId(Long id);
     List<Paciente> consultarTodos();
     Paciente guardar(Paciente paciente);
     Paciente modificar(Paciente paciente);
-    void eliminar(Integer id);
+    void eliminar(Long id);
 }
